photomap-core-api/photo-api-python/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
flask-restx==1.3.0
PyJWT==2.10.1
psycopg2-binary==2.9.10
SQLAlchemy==2.0.41
alembic==1.16.4
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
pydantic==2.11.7
Pillow==11.3.0
google-cloud-storage==3.2.0
kafka-python==2.2.15
requests==2.32.4
python-dotenv==1.1.1
gunicorn==23.0.0
imagehash==4.3.2
pytest==8.4.1
pytest-flask==1.3.0
python-multipart==0.0.20
Werkzeug==3.1.3
dependency-injector==4.48.1
flask-openapi3==2.1.3